Introduce an object-oriented C++ API based on a new C ABI
The current C API prevents adding new features and the introduction of an object oriented C++ interface. Every change results directly in a ABI break. To prevent this a new C API is introduced that is easier to extend and to maintain. On top of this API are 2 C++ classes: jpeg_encoder and jpeg_decoder to encode and decode JPEG-LS files.

This commit comes also with the following changes:
- Updated .clang-format
- Switch to Visual Studio 2019 for the CI build pipeline
- Support for reading and writing SPIFF header
- Incomplete support for JFIF header has been removed
- Extended unit test coverage

+++ src/lookup_table.h
@@ -2,12 +2,11 @@
 
 #pragma once
 
-#include <cstring>
 #include <array>
 #include <cassert>
+#include <cstring>
 
-namespace charls
-{
+namespace charls {
 
 // Tables for fast decoding of short Golomb Codes.
 struct Code final
@@ -50,7 +49,7 @@ public:
         const int32_t length = c.GetLength();
         ASSERT(static_cast<size_t>(length) <= byte_bit_count);
 
-        for (int32_t i = 0; i < 1 << (byte_bit_count - length); ++i)
+        for (size_t i = 0; i < static_cast<size_t>(1U) << (byte_bit_count - length); ++i)
         {
             ASSERT(types_[(static_cast<size_t>(value) << (byte_bit_count - length)) + i].GetLength() == 0);
             types_[(static_cast<size_t>(value) << (byte_bit_count - length)) + i] = c;
